fix: do not install "log" file outputs from pure manifest builds
It's not currently possible to install a package published from a pure
sandbox manifest build on account of the fact that a) Flox currently
installs all outputs by default and b) pure sandbox manifest builds
include a "log" output which is a file, and therefore not able to be
installed to an environment package containing symlinks to other packages.

This patch addresses the problem by disregarding "log" outputs pointing
to files, thus restoring the ability to install packages published from
pure sandbox builds. These lines should be removed once we add full
support for `outputsToInstall`, at which point skipping over the "log"
output will happen naturally as it will not be included in that list.