fix: upgrade by iid in separate group with other packages in 'toplevel'
When upgrading by install id, we check
if there are any other packages in the 'toplevel' group.
If so the upgrade fails with

```
flox upgrade my-pkg
❌ ERROR: 'my-pkg' is a package in the group 'toplevel' with multiple packages.
To upgrade the group, specify the group name:
    $ flox upgrade toplevel
To upgrade all packages, run:
    $ flox upgrade
```

However, we never check whether the upgraded package
actually belongs to the 'toplevel' group itself.
Thus installing a flake, or a catalog package into
with a different 'pkg-group' also causes this error.

```
> flox list -c
version = 1

[install]
hello.pkg-path = "hello"
htop.pkg-path = "htop"
htop.pkg-group = "not-toplevel"
flox.flake = "github:flox/flox"

> flox upgrade htop
❌ ERROR: 'htop' is a package in the group 'toplevel' with multiple packages.
To upgrade the group, specify the group name:
    $ flox upgrade toplevel
To upgrade all packages, run:
    $ flox upgrade

> flox upgrade flox
❌ ERROR: 'flox' is a package in the group 'toplevel' with multiple packages.
To upgrade the group, specify the group name:
    $ flox upgrade toplevel
To upgrade all packages, run:
    $ flox upgrade
```

This commit adds a second condition to
`pkg_belongs_to_non_empty_toplevel_group`
that requires the given install id to _also_ be member
of the 'toplevel' group.