chore(containerize): Disable tests aarch64-darwin
These are failing very frequently at the moment, which is slowing down
work on other tickets, and isn't helping us understand or fix the
underlying problem any quicker. We've prioritised the ticket for this
coming sprint so temporarily disable the tests in the meantime.

I originally tried adding this to `matrix.exclude` in the CI config but
it would require changes to disable required checks in GitHub which I
don't have access to and we might forget to re-enable.