test: regression test for spaces in PATH
The 1.3.13 release contains the new path helper that's written in Rust.
There is a bug in how this is called that does not handle spaces in
PATH. This bug wasn't caught in the test suite, so this change adds a
new collection of tests meant to catch this regression in the future.