fix: spaces in PATH for fish shell
There are actually two bugs fixed by this change:
- In-place activations for fish shells would use the generation-time
  PATH instead of the runtime PATH, meaning that a user's RC-file
  config could be removed from PATH.
- PATH that contains a space gets split in the Bash shell that
  generates the startup commands for the fish shell. When the fish
  shell then calls `flox-activations fix-paths` it sees too many
  arguments due to the word splitting that happened in Bash.

This change addresses the bugs by properly quoting in the Bash shell.
One bit of weirdness involved is that fish doesn't have variable
expansion syntax like Bash does, so there's no built-in way to expand
a variable with a default value if it's unset. To address that we use a
temporary variable that we manually expand and set to a default value.