fix(test): avoid race condition at first parallel run of podman
Check that podman is functioning
and ensure it has created the necessary directories.
Without this, starting multiple podman containers in parallel,
may cause a race between the containers to create the directories.

We started seeing the effect of the race condition frequently
when running the containerize.bats tests, both locally and in CI [^1].
When running `flox containerize` (which by default writes into a registry)
was run on macos, two processes of `podman` are spawned;

1. `podman run ...` which runs `flox containerize` from within a container and writes that to stdout
2. `podman import` which adds the created container to the local runtime.

As part of their initialization both connect via ssh to the "podman machine",
a VM specifically for running containers.
During this connection both processes probe `$HOME/.ssh` and create it if is absent [^2].
However this leaves a window where both processes see the directory as absent
and consequently attempt to create it.
If the race is "lost" by the `podman run` process, it fails to start up with

    Error: unable to connect to Podman socket: mkdir /tmp/home.AywAp0/.ssh: file exists

(Assumption) If the race is lost by the `podman import` the process appears to silently die,
causing subsequent writes to its `stdin` to err `BrokenPipe`.
However, these pipe errors are communicated via a SIGPIPE signal, that we chose to ignore
to avoid a panic when printing to `stdout` with `stdout` closed [^3] [^4].

Somewhat tellingly,

    flox show hello  | true

returns with a 141 exit code, exactly as the containerize command.

The silver lining of these issues is that they are unlikely to affect users,
unless you frequently run `containerize` on clean macos machines,
without e.g. `$HOME/.ssh` directories.

[^1]: <https://github.com/flox/flox/issues/2743>
[^2]: <https://github.com/containers/common/blob/3c510166fd8563d1c4da81aba3c6d5c7a9cc0b9b/pkg/ssh/connection_golang.go#L316-L320>
[^3]: <https://github.com/flox/flox/blob/2c404edf5ed0503cb1c366fc6c735d64d26088b1/cli/flox/src/main.rs#L51-L52>
[^4]: <https://github.com/flox/flox/issues/1193>