refactor(activate): Move FLOX_SHELL to getopt
This was broken for in-place activations by a preceding commit because
we took `rendered_exports` after adding `FLOX_SHELL` in `activate.rs`
when it shouldn't be propagated into the activation.

Fix it by passing the value with `getopt` so that we don't have to do
special handling to undo it in `activate.rs`. We still need to unset in
`activate` though because the parent shell calling `flox activate` may
have exported it and we don't want it to be carried into the next `flox
activate` call

This was previously tested for command activations but not for in-place
activations, which behave differently because we render set/unset
commands out to the current shell. These tests are pretty noisy when
they fail because they dump all of the environment but the alternatives
don't feel great:

- `--print-script` might not behave quite the same as an actual `eval`
- pre-parsing with `| grep` gets complicated for the command activation
  due to how Bats `run` works

Tests that run the `activate` script directly needed adjusting because
`FLOX_SHELL` no longer has any effect. I don't believe we've ever
advertised this as a feature so we can consider it a non-breaking
change.

It feels safer to use the recently added `_FLOX_SHELL_FORCE` at the
beginning of these tests than add `--shell` arguments to each call, in
case we forget to add one in the future and the only feedback we get is
from the `refute_output` assertion when it loads the wrong RC file.