docs(activate): Document FLOX_ENV_DESCRIPTION
Since we already mention it in `manifest.toml` examples and we now guard
for it being correct in nested activations.

This environment variable isn't currently set for builds or containers
but they don't use `flox activate` (they call the `$FLOX_ENV/activate`
directly) so it doesn't seem appropriate to describe that in this man
page.