bump: 1.3.14 -> 1.3.15
1.3.15 was tagged from the same commit as 1.3.14 because it only
included changes to `flox/flox-installers` but we forgot to bump
`VERSION`. This didn't affect the installers and doesn't warrant a new
release because it is overridden by the installers.

It will mean that flake installs of the tag report the incorrect
version (and e.g. `flox containerize` will use the wrong proxy
container) but they aren't as well supported or documented as the
installers.