fix: install correct nodejs package
We recently fixed the versions reported by catalog server, which allows
us to install Node.js packages by proper semver versions. The tricky
part is that Node.js major versions are separate packages. This change
makes it such that a user that has a version requirement in their
`package.json` will get the correct `nodejs_*` package installed that
corresponds to their version requirement.

This is accomplished by first querying the available node versions in
the catalog, then issuing a resolution request with separate package
groups for each package to see which ones match the semver requirement.
The latest version that matches is used as the package we install for
the user.