fix: install correct nodejs package
We recently fixed the versions reported by catalog server, which allows
us to install Node.js packages by proper semver versions. The tricky
part is that Node.js major versions are separate packages. This change
makes it such that a user that has a version requirement in their
`package.json` will get the correct `nodejs_*` package installed that
corresponds to their version requirement.

This is accomplished by doing parallel resolutions against a list of
hardcoded attribute paths for all currently available Node.js versions
in the catalog. This means that we'll have to add attribute paths to
this list as they become available, but that's a problem for later.

Since the current spinner was only designed for a single catalog
request at a time, you temporarily see multiple spinners (one for each
outstanding request). I've added a new spinner that says "searching for
compatible node versions" so that there is context for the multiple
spinners.