fix(containerize): Re-use cache across versions
This takes us back to using a single `flox-nix` cache volume so that
users don't suffer complete cache invalidations or have to manually
garbage collect volumes after each upgrade of Flox.

We'll mention in the release notes that the version suffixed volumes
that were created between v1.3.12 and v1.3.15 (inclusive) can be deleted
after upgrading.

This is achieved by starting a container before creating the
`ContainerSource` command which copies the contents of the container
image's store into the cache volume mounted at a different location. The
copy is idempotent, so it's cheaper if the cache volume already has the
necessary store paths, unlike `nix-store --export | nix-store --import`.

Running this command early also has the side-effect of adding an
additional guard against the race condition seen in fb881cfc.

Some notes about the examples below:

- PATH has to be overridden to favour Docker until we fix 2538
- container images are already cached so pull times are not shown
- my broadband isn't super fast so substitution still takes a little

Using the Flox from before we adding version suffixing to the volume:

    % PATH="$(dirname $(which docker)):$PATH" nix run github:flox/flox/v1.3.11 -- containerize -d ~/tmp
    ✨ 'tmp:latest' written to Docker runtime

    (2) ~/projects/flox/flox on dcarley/2661-containerize-cache-pt200 [$>] took 44s
    % PATH="$(dirname $(which docker)):$PATH" nix run github:flox/flox/v1.3.11 -- containerize -d ~/tmp
    ✨ 'tmp:latest' written to Docker runtime

    (2) ~/projects/flox/flox on dcarley/2661-containerize-cache-pt200 [$>] took 10s

Using the new caching strategy on top of that existing `flox-nix `volume
to demonstrate compatibility. It's slightly slower initially because we
no longer have Flox store paths from the container image:

    % PATH="$(dirname $(which docker)):$PATH" FLOX_CONTAINERIZE_FLAKE_REF_OR_REV=v1.3.15 flox containerize -d ~/tmp
    ✨ 'tmp:latest' written to Docker runtime

    (2) ~/projects/flox/flox on dcarley/2661-containerize-cache-pt200 [$>] took 1m38s
    % PATH="$(dirname $(which docker)):$PATH" FLOX_CONTAINERIZE_FLAKE_REF_OR_REV=v1.3.15 flox containerize -d ~/tmp
    ✨ 'tmp:latest' written to Docker runtime

    (2) ~/projects/flox/flox on dcarley/2661-containerize-cache-pt200 [$>] took 16s

Using a newer `nixos/nix` container and the same cache volume again:

    % PATH="$(dirname $(which docker)):$PATH" FLOX_CONTAINERIZE_FLAKE_REF_OR_REV=v1.3.15 FLOX_CONTAINERIZE_PROXY_IMAGE_REF=2.24.12 flox containerize -d ~/tmp
    ✨ 'tmp:latest' written to Docker runtime

    (2) ~/projects/flox/flox on dcarley/2661-containerize-cache-pt200 [$>] took 16s

Using an older version of Flox where we changed a lot of dependencies
and we have to fetch them from substitutors, which should still be
slightly faster than pulling a new Flox image:

    % PATH="$(dirname $(which docker)):$PATH" FLOX_CONTAINERIZE_FLAKE_REF_OR_REV=v1.3.14 flox containerize -d ~/tmp
    ✨ 'tmp:latest' written to Docker runtime

    (2) ~/projects/flox/flox on dcarley/2661-containerize-cache-pt200 [$>] took 1m41s
    % PATH="$(dirname $(which docker)):$PATH" FLOX_CONTAINERIZE_FLAKE_REF_OR_REV=v1.3.14 flox containerize -d ~/tmp
    ✨ 'tmp:latest' written to Docker runtime

    (2) ~/projects/flox/flox on dcarley/2661-containerize-cache-pt200 [$>] took 17s