feat(upgrade): Reference dry-run in notifications
To help users see what upgrades are available before committing to them.
Per the design in 2572 (which might have been missed due to the ordering
and size of the tickets) and the suggestion from Eric Torreborre in this
thread:

- https://floxcommunitygroup.slack.com/archives/C06NYALM8P8/p1740482265759719

Also updates all occurrences of the message in docs to match.