fix(containerize): Set FLOX_ENV_{DESC,CACHE}
Follow-up from 2178ba3e for the documented environment variables that
are now passed as args and can be derived for containers. Values are as
follows..

`--env-project`: not set because there is no source project in the
rendered container. So `FLOX_ENV_PROJECT` will remain unset.

`--env-cache`: root of `/tmp` since we're in this middle ground of not
having a `.flox` but we're still (currently) running in `dev` mode and
users may wish to make use of `PROJECT_ENV_CACHE`.

`--env-description`: bare environment name which matches the behaviour
of path environments in the CLI, since a remote or managed environment
no longer has any parent information when it's rendered to a container.