test(containerize): Remove FLAKE_REF_OR_REV
Use the matching version of Flox in the proxy container on macOS now
that we are reporting versions correctly (or nearly, after 2774).

This will cause failures if you run the tests locally without pushing
the current HEAD to a GitHub branch but that's a clearer failure than
the current behaviour where the tests will behave unexpectedly on macOS
if you make changes to anything that happens within the proxy container
such as `mkContainer`.