fix(activate): Don't propagate FLOX_VERSION
I noticed that `flox --version` and the `containerize` tests were still
incorrect for me when the devShell is running from within an interactive
activation:

    % env -i /run/current-system/sw/bin/flox --version
    1.3.14-g8fa0811
    % ./cli/target/debug/flox --version
    1.3.14-g8fa0811
    % env -i ./cli/target/debug/flox --version
    1.3.15-g2c0ae4c3

We can't easily test the wrapper script from the integration tests so
they attempt to exercise the same combination of code paths. Failure
from before this change:

    version.bats
     ✓ version: accepts runtime version from wrapper derivation [917]
     ✗ version: doesn't propagate runtime version into activations [1147]
       tags: version
       (from function `assert_output' in file /nix/store/yf08fkff6pbr4lya678aprc0nwsddxgn-bats-with-libraries-1.11.1/share/bats/bats-assert/src/assert_output.bash, line 178,
        from function `assert_buildtime_version' in file version.bats, line 58,
        in test file version.bats, line 76)
         `assert_buildtime_version' failed
       ✨ Created environment 'project-2' (aarch64-darwin)

       Next:
         $ flox search <package>    <- Search for a package
         $ flox install <package>   <- Install a package into an environment
         $ flox activate            <- Enter the environment
         $ flox edit                <- Add environment variables and shell hooks

       1.2.3

       -- regular expression does not match output --
       regexp : [0-9]+\.[0-9]+\.[0-9]+-g.+
       output : 1.2.3
       --

       Last output:
       1.2.3
     ✓ version: uses buildtime version in absence of wrapper derivation [894]

Fix this by eagerly evaluating the version when `flox` starts and then
unsetting the `FLOX_VERSION` environment variable once we've read it. I
considered putting the unset in `FLOX_VERSION_STRING` but it feels
clearer to do it at the point we evaluate. Matthew kinda predicted this
in:

- https://github.com/flox/flox/pull/2659/files#r1934753653

I was going to convert the environment variable name to a constant but
`option_env!` requires a string literal.