feat: add composition to lockfile
This adds an optional "compose" section to the lockfile. When this
section is present it indicates that the lockfile's manifest was the
result of merging other manifests.

An alternative to making this section optional would be to make the
`compose.manifest` field optional. But that means we have a `compose`
section in the lockfile even when we haven't done composition.