refactor: make include name optional
The name of an included environment is stored as part of the
environment, but may collide with other names or just not be the
desired name. Making this field optional means that we can infer the
name for a user, while still letting them override it if they choose.

Storing the concrete name in the lockfile allows us to persist the name
because it's not stored in the manifest of a path environment, and
would otherwise not exist in the lockfile.