fix: prevent panic during upgrade
In some cases you may encounter a package that was in the old lockfile
and isn't in the new lockfile (or isn't present for a certain system).
We've encountered this in production, which most likely means that the
manifest that was present when initiating the upgrade check was out of
sync with its lockfile (i.e. someone edited the manifest through means
other than `flox edit`). In those cases we silently ignore the packages
(or packages for a certain system) that are no longer present.