feat(activate): Add activate.mode to manifest
This provides a default for the environment which still defaults to
`dev` if not specified and can still be overridden by `flox activate -m`

The `manifest.toml` docs have been copied and slightly modified from:

- https://flox.dev/docs/concepts/activation/#development-vs-runtime-mode

The `activate` docs have been slimmed down with a reference to
`manifest.toml` so that we don't have to repeat all of the same
information; some of which already varied between the three places they
were specified. I plan to change the concepts do to link back to
`manifest.toml` docs as well.

The new `[activate]` section of the manifest has been added before
`[options]` and not included in the manifest template to reflect that
most users won't need to set it and it's not a first-class feature that
we need to draw attention to like `[services]` are.