feat(containerize): Use manifest activate.mode
Make `flox containerize` honour the `activate.mode` setting in the
manifest rather than always defaulting to the rendered `dev`
environment.

We don't need to do anything additional for macOS because the project
path (parent of `.flox`) is passed over to the proxy container and it
re-uses the manifest.

I haven't added tests for this because it's pretty simple and the
containerize tests are a pain to work with, especially on macOS.

This could also be overridden with a flag to `flox containerize` like
`flox activate -m run` does but it needs plumbing all the way through
for macOS and again testing is hard, so we can follow-up on that later.