feat(init): Manifest mode=run for default envs
Set the new manifest option when creating a default environment, either
through `flox init -d ~` or `flox install` outside a project. This
reduces the chance of accidentally mixing activation modes either when
adding it to your RC files or additional manual activations.

I'm leaving `-m run` in the suggestions and docs for consistency with
older versions, in case someone uses the new docs without upgrading to
the most recent version.