fix: add host attr of github/gitlab flakerefs to URL serialization
Backport of upstream PR targeting nix >= 2.27
<https://github.com/NixOS/nix/pull/12580>

When installing `github` or `gitlab` flake refs with an alternative host to an environment,
the environment will fail to build in the general case where the store paths are not substitutable:

```
flox install gitlab:foo/bar?host=gitlab.example.com

❌ ERROR: Failed to build environment:

Failed to realise 'bar':
error:
       … while fetching the input 'gitlab:foo/bar/87687do78you0876ear1eallya2eacd62read77cadthis'

       error: unable to download 'https://gitlab.com/api/v4/projects/foo%2bar/repository/archive.tar.gz?sha= 87687do78you0876ear1eallya2eacd62read77cadthis': HTTP error 404

       response body:

       {"message":"404 Project Not Found"}
```

That is because nix drops the `host` parameter when serializing the locked flakeref to a url.
The included patch adds the `?host` query back to the locked url.