refactor: rename LockedManifestError -> ResolveError
Locking now includes merging, which will have to return an
EnvironmentError, because it will have to open both path and managed
environments. So locking will have to return an EnvironmentError.

Almost all the current errors on LockedManifestError are resolution
errors, and we should have consistency between the name of the resolve
operation and the name of the error it returns. So rename
LockedManifestError -> ResolveError.

Only two errors on LockedManifestError aren't resolution errors, so move
those up to CoreEnvironment. This seems to make sense anyways, because
it removes `.map_err(CoreEnvironmentError::LockedManifest)?;` in
`CoreEnvironment`.