refactor: return EnvironmentError everywhere we lock
Upcoming merge functionality will have to return an EnvironmentError,
since merging environments will have to call ManagedEnvironment::open
and PathEnvironment::open which return EnvironmentError and
ManagedEnvironmentError.

Merging has to occur between operations currently implemented in
core_environment. E.g. for install we'll have to
1. insert_packages
2. merge
3. resolve
4. build

It's possible it could make more sense to merge outside of
core_environment, but if we decide to do that, it would require a larger
refactor, since if we merge outside of core_environment, we'd have to
trigger steps 3 and 4 outside of core_environment as well.

It seems much more straightforward to just merge within
core_environment, which means all operations that lock will have to
return an EnvironmentError.

We can keep EnvironmentError, CoreEnvironmentError, and
ManagedEnvironmentError as organizational buckets, but we can't expect
to always return those errors in their respective environment,
core_environment, and managed_environment modules.