refactor: move search functionality to inner fn
This change moves the actual search functionality to an inner `search`
method that doesn't apply a spinner. The `search_with_spinner` method
now just calls the `search` method with the spinner applied.

The `search_with_spinner` call in the node setup is now replaced with
the `search` function so that "search for packages that match 'nodejs_'"
isn't shown to users during that operation.