feat: match yarn version to yarn-berry during init
The yarn package manager did a rewrite after version 1. In the larger
ecosystem it seems that the new package is still called "yarn", but in
nixpkgs it is now called "yarn-berry" because "berry" was the name of
the rewrite project. The auto-setup code was skipping yarn installation
because there is no package in the catalog called "yarn" with a version
>=2.