feat: IncludeFetcher
Add the IncludeFetcher struct, and thread it throughout the rest of the
codebase.

Path/Managed/Remote environments need to configure our merge code to
specify whether or not local includes should be allowed, and they need
to configure the directory relative to which includes should be fetched.

Use IncludeFetcher to accomplish this.

This required returning EnvironmentError in a few places in
managed_environment in order to return EnvironmentError::InvalidPath.
Since that error is already returned by ManagedEnvironment::parent_path,
it doesn't make sense to return a different error for what is
essentially the same thing.