feat(compose): Add support to list --config
Update `flox list --config` to print the merged manifest with a notice
when using composition.

We now lock early, which we didn't previously for `list -c`, in order to
detect composition and for a more consistent experience between composed
and non-composed environments.

We have to re-serialize merged manifests from `typed::Manifest` which
means that we lose comments and add some defaults that the user may not
have specified. We'll improve on this in later commits and tickets.