fix: use lockfile to detect changes for edit
Currently, if `flox edit` modifies `[include]` in a way that leads to a
change in e.g. hook.on-activate for the merged manifest, we don't print
the warning that a user may need to reactivate.

Fix this by using the manifest in the lockfile rather than the user
manifest.

This also refactors EditResult to only have two variants, Unchanged and
Changed. Upcoming changes will require warning when `[include]` is
modified, which isn't mutually exclusive with ReActivateRequired and
Success, so the enum would start to become a truth table. Using a method
EditResult::reactivate_required will be simpler.