fix(installer_channel): Document and restrict vals
We had a Sentry report for a failed update check against the path
`by-env/nixos-24.05/LATEST_VERSION` which looks like someone set the
`installer_channel` config option to represent a Nix channel:

- https://flox-dev.sentry.io/issues/6421529049/?alert_rule_id=15667377&alert_type=issue&environment=stable&notification_uuid=cb2bb2c3-3b62-4c88-b17a-fb67d0426310&project=4506548241825792&referrer=slack

Prevent this from happening by restricting and documenting the values
that can be provided. If you try to set an incorrect value imperatively:

    % flox config --set installer_channel nixos-24.05
    ❌ ERROR: unknown variant `nixos-24.05`, expected one of `stable`, `nightly`, `qa`

If you set an incorrect value declaratively, as that user has:

    % flox list
    ❌ ERROR: Could not parse config: unknown variant `nixos-24.05`, expected one of `stable`, `nightly`, `qa`

Unfortunately the errors don't include the path of the invalid key but
in this case it should be clear enough where the problem lies. I looked
briefly at whether `serde_path_to_error` would help here but it doesn't
appear to give anything for errors in top-level keys like this.