test(integ): Default to FLOX_FEATURES_COMPOSE=true
Run all integration tests with the composition feature flag enabled so
that we get early feedback in preparation for turning it on for
everywhere when the feature is launched.

Ideally it would be nice to run a matrix of the tests with and without
the feature flag but it's complicated to configure and expensive to run,
so it doesn't seem worth it.

It's worth noting that the feature flags for composition and services
affect manifest parsing in various places and we mostly switch on the
presence of includes rather than the flag, unlike pkgdb vs catalog which
took completely different setups and codepaths.

This change highlights an issue with `flox list` which will be fixed in
a subsequent commit.