fix: remove unwraps in init
We recently received a Sentry panic with:
```
called `Result::unwrap()` on an `Err` value: Os { code: 2, kind: NotFound, message: "No such file or directory" }
```

It happened within the `flox::commands::init::init` trace.

`current_dir()` returns an error if the current directory does not
exist, so that seems like a likely cause for the error.

Even if that wasn't the cause, we should add some context in case the
error happens again.