test(node): Fix mocks for Yarn 1.x
This was incorrectly resolving to `yarn-berry` after 3dca25e7 because
there was no version constraint in the `package.json` being used.

However it only surfaced when the mocks were refreshed by a catalog
client update which failed the integration test because the `yarn.lock`
format didn't match the intended version:

    # ➤ YN0028: The lockfile would have been modified by this install, which is explicitly forbidden.
    # ➤ YN0000: └ Completed
    # ➤ YN0000: · Failed with errors in 0s 167ms
    # Internal Error: single-dependency@workspace:.: This package doesn't seem to be present in your lockfile; run "yarn install" to update the lockfile

Fix this by using a separate `package.json` that specifies the older
branch of the Yarn package. The mock has been regenerated locally on my
machine so that systems order differs to before.