test(node): Merge common dir into npm
This `package.json` is only used by the test for `npm` now. All other
variants need their own `package.json` because they need to specify
different `engines`.

We now have a few copies of the `index.js` app and not all of the tests
run an activation with the app but I don't have strong feelings about
duplicating that small amount of content for easier to read tests and
slowing down the other tests by making them all run activations.