feat: add zebra() to Environment trait
We haven't yet decided what to call the operation that re-fetched an
included environment, so I'm calling that zebra in this commit.

Add zebra() to Environment trait and implement it in CoreEnvironment and
Lockfile.

zebra() re-fetches all included environments or if specified, a list of
included environments.

I added some tests, but I plan to integration test the following after
this functionality is exposed with a flox subcommand:
- zebraing all included environments
- zebraing by name (and making sure nothing else is zebraed)
- zebraing by directory name