feat(compose): fn for printing overridden fields
This will be wired into sub-commands in a subsequent commit. Those
integration tests will contain simpler set of data because they require
more boilerplate to setup.

The original iteration of this preserved the merge order of the
environments from the order of the warnings but was lost when we started
de-duplicating the fields by environment.

We can bring that back later if we decide that it's important but we'll
need to decide whether we trust the order of the warnings or the order
of `Compose.includes`, since there's no guarantee that they contain the
same names or order.