feat(compose): Print overrides for edit and list
This isn't currently implemented for other commands that may lock like
`activate` because we need to surface when an environment is being
freshly locked vs already locked.

I've matched the locations of the existing "integration" tests in BATS
and Rust because both commands use a different pattern and I still can't
decide which I prefer.

This surfaces the name `Current manifest` which isn't ideal but we don't
currently have a way for the `ComposesiteManifest` to know the name of
the environment that it came from.