test(services): Add activate_finished fifo
Block this test on the activation having started (which in turn blocks
on `process-compose` starting) before attempting to poll for the running.
This pattern is copied from tests in `activate.bats`.

To workaround these failures on the x86_64-darwin remote builder:

    not ok 472 activate services: shows warning when services already running in 2105ms
    # tags: services
    # (in test file services.bats, line 938)
    #   `"${TESTS_DIR}"/services/wait_for_service_status.sh one:Running' failed with status 124
    # ✨ Created environment 'test' (x86_64-darwin)
    #
    # Next:
    #   $ flox search <package>    <- Search for a package
    #   $ flox install <package>   <- Install a package into an environment
    #   $ flox activate            <- Enter the environment
    #   $ flox edit                <- Add environment variables and shell hooks
    # ✅ Environment successfully updated.
    # Service 'one' status current=Stopped, expected=Running
    #
    # Last output:
    # ✅ Environment successfully updated.

Recent examples:

- https://github.com/flox/flox/actions/runs/13639271677/job/38125290933
- https://github.com/flox/flox/actions/runs/13944603552/job/39028991616
- https://github.com/flox/flox/actions/runs/13657894699/job/38182980609
- https://github.com/flox/flox/actions/runs/13636847423/job/38117625355

What's particularly strange about the failures is that it only completed
one or two polls of loop (with 0.1s delay) within the 1s timeout. The
only explanation I can think for this is that `flox services status`
blocked when calling `process-compose` because its server was slow or
failed to respond to its client.

I've been unable to reproduced this on the builder over SSH or by
introducing delays or load to my local system though.