refactor: make allow.licenses optional
This change makes `allow.licenses` optional so that its merge semantics
are clearer. This field is a `Vec`, which means that it defaults to
empty. At merge time it was unclear whether an empty `Vec` was present
as a default vs. explicitly set by the user, so we've made this more
explicit by making the field optional.

Since this field is optional, it will no longer be included in the
lockfile by default. This means that the test data needed to be
regenerated.