feat: skip serializing options.{allow,semver}
This changes omits `options.allow` and `options.semver` when their
sub-fields are `None`. This prevents `options.allow` and
`options.semver` from appearing in the merged manifest when they weren't
present in any of the composed manifests.

The reason they appeared in the first place is that they were
unconditionally serialized into the lockfile, so a manifest taken from a
lockfile would always contain these tables even though they were empty.

Since this changes the contents of the lockfile the test data needed to
be regenerated.