feat(compose): Order override notices by env
List the notices about overridden fields in the order that the
environments were included and the composing environment last.

Changes the unit test to use real environments because it depends more
heavily on the implementation of the `Compose` object; specifically the
naming of the environments in `Compose.include` vs `Compose.warnings`.