chore: check that pushed  environments do not refer to local includes
Local includes be it by absolute or relative path are unlikely to impossible to reproduce when pulling the environment on another machine.

fixup! chore: check that pushed environments do not refer to local includes

Co-authored-by: Dan Carley <dan.carley@gmail.com>

fixup! chore: check that pushed environments do not refer to local includes

Move check to Manifest/push implementation