refactor: CoreEnv.ensure_locked() -> LockResult
Adds a `LockResult` enum which indicates whether an environment was
newly locked or an existing lock was used, matching the pattern of
`EditResult` and `UpgradeResult`.

It prevents callers like `activate` needing to `lockfile_up_to_date()` -
which I'll move off the `Environment` trait in a subsequent commit -
before calling `lock()` or `ensure_locked()`.

Into (via From) is provided for accessing the inner `Lockfile` if the
caller doesn't care about the operation. I wasn't sure whether to call
this `to_inner()`, `lockfile()`, or provide it on a struct, but settled
on this.

I'm not currently making any attempt to refactor `ensure_locked()` in
relation to the (slightly out of date) comments because it requires some
additional benchmarking and it seems alright to me at the moment, but I
don't have enough new information to warrant updating the comment.