refactor(tests): Split [un]install.bats
So that the integration test file names match the sub-commands that
they're testing, which makes it easier to see at a glance how much test
coverage we have and where new tests should be located.

The only change to the tests themselves is standardising the test name
prefixes with `uninstall:` instead of `'flox uninstall'`.