feat(compose): Add errors and notices to uninstall
An imperative uninstall command on a composed environment only operates
on the composing environment's manifest rather than the merged
manifest (which is essentially a read-only view) or the included
environments (which are also read-only locks).

So return an error if you attempt to uninstall a package that is only
provided by an included environment.

Also print a notice if you uninstall a package from the composing
environment but the package is still provided by an include, which could
be surprising to the user if we said it was removed but it's still
visible.

Only the highest priority include is returned, rather than listing all
of the possible matches, for user-facing simplicity and to match how we
handle warnings for overridden manifest fields. This could result in
whack-a-mole where you have to remove a package from multiple
environments but it seems like the less common scenario.

It's worth noting that we have to lock in order to detect composition
and this could potentially cause the environment to be built twice if it
was unlocked prior running `flox uninstall`.