fix: reword include upgrade language
Change the phrasing for include upgrades to make it clear that we're
upgrading the composer environment by getting the latest changes to its
included environments, but we're not performing a `flox upgrade` on the
included environments.

Also, in the case `flox upgrade include dep1 dep2` where only one of the
included environments has changes, warn that the other environment does
not have changes.